Mastodon_crawler/requirements.txt
couchdb==1.2
Mastodon.py==1.8.1
nltk